crypto-bigint: `const fn` impls of `concat` and `split` (#111)
